Add files via upload
Jupyter Notebook containing detailed explanations of the code used to condition and prepare the data aswell as the code to train the model. Model training, learning rate adjustments, saving the model, loss determination is still a work in progress. Tasks completed: Conditionaing and Preparing the data for training the ML model. Dated: 15/09/2024